requirements.txt
# Runtime — only one of these is needed depending on --async flag:
requests>=2.28.0       # sync client (default)
httpx>=0.27.0          # async client (--async flag)

# Web UI
flask>=3.0.0

# Optional — only needed for .yaml / .yml input:
pyyaml>=6.0
